Separation of cancel and deletion privileges
Including the name of the user who has cancelled a booking and the date and time of it

+++ module/Backend/src/Backend/Controller/BookingController.php
@@ -69,7 +69,7 @@ class BookingController extends AbstractActionController
     {
         $sessionUser = $this->authorize('admin.booking, calendar.see-data');
 
-        $params = $this->backendBookingDetermineParams();
+        $params = $this->backendBookingDetermineParams(true);
 
         if (! $this->getRequest()->isPost()) {
             switch (count($params['reservations'])) {
@@ -287,7 +287,9 @@ class BookingController extends AbstractActionController
 
     public function deleteAction()
     {
-        $this->authorize('admin.booking, calendar.cancel-single-bookings, calendar.cancel-subscription-bookings');
+        $sessionUser = $this->authorize(['admin.booking',
+            'calendar.cancel-single-bookings', 'calendar.delete-single-bookings',
+            'calendar.cancel-subscription-bookings', 'calendar.delete-subscription-bookings']);
 
         $serviceManager = $this->getServiceLocator();
         $bookingManager = $serviceManager->get('Booking\Manager\BookingManager');
@@ -301,16 +303,17 @@ class BookingController extends AbstractActionController
 
         switch ($booking->get('status')) {
             case 'single':
-                $this->authorize('admin.booking, calendar.cancel-single-bookings');
+                $this->authorize(['admin.booking', 'calendar.cancel-single-bookings', 'calendar.delete-single-bookings']);
                 break;
             case 'subscription':
-                $this->authorize('admin.booking, calendar.cancel-subscription-bookings');
+                $this->authorize(['admin.booking', 'calendar.cancel-subscription-bookings', 'calendar.delete-subscription-bookings']);
                 break;
         }
 
         if ($this->params()->fromQuery('confirmed') == 'true') {
 
             if ($editMode == 'reservation') {
+                $this->authorize(['admin.booking', 'calendar.delete-single-bookings', 'calendar.delete-subscription-bookings']);
 
                 $reservationManager->delete($reservation);
 
@@ -318,11 +321,17 @@ class BookingController extends AbstractActionController
             } else {
 
                 if ($this->params()->fromQuery('cancel') == 'true') {
+                    $this->authorize(['admin.booking', 'calendar.cancel-single-bookings', 'calendar.cancel-subscription-bookings']);
+
                     $booking->set('status', 'cancelled');
+                    $booking->setMeta('cancellor', $sessionUser->get('alias'));
+                    $booking->setMeta('cancelled', date('Y-m-d H:i:s'));
                     $bookingManager->save($booking);
 
                     $this->flashMessenger()->addSuccessMessage('Booking has been cancelled');
                 } else {
+                    $this->authorize(['admin.booking', 'calendar.delete-single-bookings', 'calendar.delete-subscription-bookings']);
+
                     $bookingManager->delete($booking);
 
                     $this->flashMessenger()->addSuccessMessage('Booking has been deleted');
